+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/package-info.java
@@ -28,6 +28,7 @@
  * <li>{@link is.codion.swing.common.ui.component.table.FilterTableCellRenderer#ALTERNATE_ROW_COLORING}
  * <li>{@link is.codion.swing.common.ui.component.table.FilterTable#COLUMN_REORDERING}
  * <li>{@link is.codion.swing.common.ui.component.table.FilterTable#AUTO_RESIZE_MODE}
+ * <li>{@link is.codion.swing.common.ui.component.table.FilterTable#COLUMN_RESIZING}
  * </ul>
  */
 package is.codion.swing.common.ui.component.table;